% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSVM.R
\name{fx_hsvm}
\alias{fx_hsvm}
\title{Objective Function for HSVM}
\usage{
fx_hsvm(y, X, beta, delta = 0.4, lambda = 1e-04)
}
\arguments{
\item{y}{binary response}

\item{X}{design matrix}

\item{beta}{regression coefficient vector}

\item{delta}{shape parameter}

\item{lambda}{regularization parameter}
}
\description{
Objective Function for HSVM
}
