% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSVM.R
\name{gradf_hsvm}
\alias{gradf_hsvm}
\title{Gradient for HSVM}
\usage{
gradf_hsvm(y, X, beta, delta = 0.4, lambda = 1e-04)
}
\arguments{
\item{y}{binary response}

\item{X}{design matrix}

\item{beta}{regression coefficient vector}

\item{delta}{shape parameter}

\item{lambda}{regularization parameter}
}
\description{
Gradient for HSVM
}
